fix(iOS):update iOS header

+++ release/iOS/x86_64/include/DataReporter.h
@@ -14,7 +14,7 @@ DATA_REPORTER_EXPORT @interface DataReporter : NSObject
     
 + (void)SetFileMaxSize:(void *)nativeReporter
            fileMaxSize:(NSUInteger)fileMaxSize;
-
+		   
 + (void)SetExpiredTime:(void *)nativeReporter
            expiredTime:(int64_t)expiredTime;
 		   
@@ -29,10 +29,10 @@ DATA_REPORTER_EXPORT @interface DataReporter : NSObject
         data:(NSString *)data;
 
 + (void)UploadSucess:(void *)nativeReporter
-                 key:(NSInteger)key;
+                 key:(int64_t)key;
     
 + (void)UploadFailed:(void *)nativeReporter
-                 key:(NSInteger)key;
+                 key:(int64_t)key;
     
 + (void)ReleaseReporter:(void *)nativeReporter;
